update: DBに格納するときに成功/失敗を分かりやすくした ユーザがないとき，学生番号が長すぎるときはクライアント側のミスとして400 それ以外のエラーはとりあえず500 特に問題ないときは200
