+++ next-app/src/app/api/v1/[studentId]/report/route.ts
@@ -22,7 +22,22 @@ export async function POST(request: NextRequest, { params }: { params: { student
   const reqBody = await request.json() as ReportPostRequest
   const studentId = params.studentId
 
-  const res = registerNewReport({...reqBody, seminarDate: new Date(reqBody.seminarDate), studentId: studentId})
+  try {
+    const res = registerNewReport({ ...reqBody, seminarDate: new Date(reqBody.seminarDate), studentId: studentId })
+    
+    if (res == "Student ID must be less than 8 characters.") {
+      return new Response(res, { status: 400 })
+    }
 
-  return Response.json(res)
+    if (res == "Author is required.") {
+      return new Response(res, { status: 400 })
+    }
+
+    return new Response(JSON.stringify(res), { status: 200 })
+
+  } catch (e) {
+    if (e instanceof Error) {
+      return new Response(e.message, { status: 500 })
+    }
+  }
 }
